Add fastapi endpoints for files and directories
